updated files
updated User.java to make sure name, username and password are consistent . Updated user repository to do a query to findAll users.  Updated the LoginControllers check username and password so that if valid its sends you to the welcome page. Updated register.html and added a redirect. html page to tell user that login is incorrect, try again.